Integrate Tmux server and Window Manager with SystemD.
Lots of work has been done in improving how we should work with systemd from the user perspective.

Update tmux service file to be run as user instead of it being run as the root user for normal user instances...

Update many configuration files to simplify or remove old code.